feat: add filters for columns and split out separate files
